% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folderfun.R
\docType{package}
\name{folderfun}
\alias{folderfun}
\title{Manages folder functions}
\description{
If you find yourself working on multiple different projects in R, you'll want
a series of folders pointing to raw data, processed data, plot results,
intermediate table outputs, etc. This package makes it easier to do that by
providing a quick and easy way to create and use functions for project-level
directories.
}
\references{
\url{http://github.com/databio/folderfun}
}
\author{
Nathan C. Sheffield
}
