% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folderfun.R
\name{inff}
\alias{inff}
\title{Prepends a parent folder to a given filename or directory.}
\usage{
inff(ffName, ...)
}
\arguments{
\item{ffName}{The name of a variable (either an R option or an environment
variable) that contains the path to the root folder to prepend}

\item{...}{One or more arguments that, when concatenated, specify a relative
path to a file or folder within the parent folder represented by the
folder function.}
}
\description{
Generic function to prepend an environment variable directory to your
relative filepath. This helps make your scripts portable because you can
point to all files with a relative path. This function uses a variable to
identify pre-specified root folders, and then it makes your relative path
into an absolute path by pre- pending the appropriate root folder. The
variables identify locations of root folders using a priority system; it look
first in R options, and then in environment variables.
}
\examples{
inff("PROC", "data.txt")
}
