% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listff.R
\name{listff}
\alias{listff}
\title{Display of project environment variables}
\usage{
listff()
}
\value{
Two-column matrix in which first column contains function 
   names and the second contains the value associated with each.
}
\description{
\code{listff} displays a two-column matrix of function names 
with which to access option / environment variable values 
associated with this package, along with the current value 
associated with each function.
}
