% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folderfun.R
\name{optOrVar}
\alias{optOrVar}
\title{Retrieval of value associated with a name}
\usage{
optOrVar(name)
}
\arguments{
\item{name}{The name of the option or env var to fetch}
}
\value{
The value associated with the given \code{name}, 
   returning \code{NULL} if and only if the \code{name} is 
   set neither as an option nor as environment variable.
}
\description{
\code{optOrVar} retrieves that value assocaited with the 
name provided as an argument, prioritizing in its search 
by first interpreting the given argument as an \code{R}
option name, and then trying an interpretation as an 
environment variable if and only if the \code{R} option 
is not set or is an empty string or vector.
}
