% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folderfun.R
\name{setff}
\alias{setff}
\title{Creates a folder function for easy access to the directory}
\usage{
setff(name, location)
}
\arguments{
\item{name}{An immutable key given to identify the current folder, and used as a
string to create the new folder function}

\item{location}{(Absolute) path to a folder that will be prepend when the
specified folder function is called}
}
\description{
Creates a folder function for easy access to the directory
}
\examples{
setff("PROC", "here")
}
